backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
python-dotenv==1.0.1
anthropic==0.43.0
